# openapi_server/models/base_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class BaseResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, code=None):  # noqa: E501
        """BaseResponse - a model defined in OpenAPI

        :param code: The code of this BaseResponse.  # noqa: E501
        :type code: str
        """
        self.openapi_types = {
            'code': str
        }

        self.attribute_map = {
            'code': 'code'
        }

        self._code = code

    @classmethod
    def from_dict(cls, dikt) -> 'BaseResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BaseResponse of this BaseResponse.  # noqa: E501
        :rtype: BaseResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def code(self):
        """Gets the code of this BaseResponse.

        响应码，0表示成功，1表示失败  # noqa: E501

        :return: The code of this BaseResponse.
        :rtype: str
        """
        return self._code

    @code.setter
    def code(self, code):
        """Sets the code of this BaseResponse.

        响应码，0表示成功，1表示失败  # noqa: E501

        :param code: The code of this BaseResponse.
        :type code: str
        """
        allowed_values = ["0", "1"]  # noqa: E501
        if code not in allowed_values:
            raise ValueError(
                "Invalid value for `code` ({0}), must be one of {1}"
                .format(code, allowed_values)
            )

        self._code = code